requirements.txt
automata_python
click